refactor: rename `constants` file to support vscode code trace
