+++ src/hooks/pageInfo/useListenLocation.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react'
-import { isLocalMode } from 'constants'
+import { isLocalMode } from 'constants/base'
 import { listenPjaxEvent } from 'utils/pjax'
 
 const listenLocation = (callback) => {
